+++ R/Codomain.R
@@ -1,12 +1,12 @@
 #' @title Codomain of Function
 #'
 #' @description
-#' A set of [Param] objects defining the codomain of a function. The parameter
+#' A [ParamSet] defining the codomain of a function. The parameter
 #' set must contain at least one target parameter tagged with `"minimize"` or
 #' `"maximize"`. The codomain may contain extra parameters which are ignored
 #' when calling the [Archive] methods `$best()`, `$nds_selection()` and
 #' `$cols_y`. This class is usually constructed internally from a
-#' [paradox::ParamSet] when [Objective] is initialized.
+#' [ParamSet] when [Objective] is initialized.
 #'
 #' @export
 #' @examples
@@ -42,55 +42,60 @@ Codomain = R6Class("Codomain", inherit = paradox::ParamSet,
     #' Creates a new instance of this [R6][R6::R6Class] class.
     #'
     #' @param params (`list()`)\cr
-    #'   List of [Param], named with their respective ID.
-    #'   Parameters are cloned.
-    initialize = function(params = named_list()) {
-      # assert parameters
-      for (param in params) {
-        # only check for codomain parameters tagged with minimize or maximize
-        if (any(c("minimize", "maximize") %in% param$tags)) {
-          # all numeric
-          if (!param$is_number) {
-            stopf("%s in codomain is not numeric", param$id)
-          }
+    #'   Named list with which to initialize the codomain.
+    #'   This argument is analogous to [ParamSet]'s `$initialize()` `params` argument.
+    initialize = function(params) {
 
-          # every parameter's tags contain at most one of 'minimize' or 'maximize'
-          if (sum(param$tags %in% c("minimize", "maximize")) > 1) {
-            stopf("%s in codomain contains a 'minimize' and 'maximize' tag", param$id)
-          }
-        }
-      }
+      assert_list(params)
+
       super$initialize(params)
 
-      # assert at least one target parameter
-      if (!any(self$is_target) && length(params)) stop("Codomain contains no parameter tagged with 'minimize' or 'maximize'")
+      # only check for codomain parameters tagged with minimize or maximize
+      for (id in self$target_ids) {
+        # all numeric
+        if (!self$is_number[id]) {
+          stopf("%s in codomain is not numeric", id)
+        }
+        # every parameter's tags contain at most one of 'minimize' or 'maximize'
+        if (sum(self$tags[[id]] %in% c("minimize", "maximize")) > 1) {
+          stopf("%s in codomain contains a 'minimize' and 'maximize' tag", id)
+        }
+      }
+
+      # assert at least one target eter
+      if (!any(self$is_target) && self$length) stop("Codomain contains no parameter tagged with 'minimize' or 'maximize'")
     }
   ),
 
   active = list(
 
     #' @field is_target (named `logical()`)\cr
-    #' Position is `TRUE` for target [Param]s.
+    #' Position is `TRUE` for target parameters.
     is_target = function() {
-      map_lgl(self$tags, has_element, "minimize") | map_lgl(self$tags, has_element, "maximize")
+      self$ids() %in% self$target_ids
     },
 
     #' @field target_length (`integer()`)\cr
-    #' Returns number of target [Param]s.
+    #' Returns number of target parameters.
     target_length = function() {
-      sum(self$is_target)
+      length(self$target_ids)
     },
 
     #' @field target_ids (`character()`)\cr
-    #' Number of contained target [Param]s.
+    #' IDs of contained target parameters.
     target_ids = function() {
-      self$ids()[self$is_target]
+      if ("any_tags" %in% names(formals(self$ids))) {
+        self$ids(any_tags = c("minimize", "maximize"))
+      } else {
+        # old paradox
+        self$ids()[map_lgl(self$tags, function(x) any(c("minimize", "maximize") %in% x))]
+      }
     },
 
     #' @field target_tags (named `list()` of `character()`)\cr
-    #' Tags of target [Param]s.
+    #' Tags of target parameters.
     target_tags = function() {
-      self$tags[self$is_target]
+      self$tags[self$target_ids]
     },
 
     #' @field maximization_to_minimization (`integer()`)\cr