+++ NEWS.md
@@ -1,3 +1,5 @@
+# bbotk (development version)
+
 # bbotk 0.8.0
 
 * fix: `OptimizerIrace` failed with logical parameters and dependencies.